requirements.txt
Django>=3.2,<4.0
djangorestframework
drf-yasg
requests
python-dotenv
pytest-django